requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
pymysql==1.1.0  # or PyMySQL if you prefer pure Python
email-validator==2.0.0
Flask-Migrate==4.0.5   # optional, for database migrations